# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend for the all rated extreme demon list",
        "title": "Aredl API",
        "contact": {
            "name": "Aredl",
            "url": "https://aredl.net/"
        },
        "version": "1.0"
    },
    "host": "api.aredl.net",
    "basePath": "/api",
    "paths": {
...
        "/aredl/names": {
            "get": {
                "description": "Gives a map of important users grouped by their role. This also includes aredl plus members",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aredl"
                ],
                "summary": "Important users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/aredl.NameUser"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/merge-requests": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Lists all open merge requests\nRequires user permission: name_change_review",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "global"
                ],
                "summary": "List name merge requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/global.NameChangeRequest"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a merge request for the user with a placeholder user. Needs to be reviewed by a moderator.\nRequires user permission: user_request_merge",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "global"
                ],
                "summary": "Merge Request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of the placeholder user to be merged with",
                        "name": "placeholder_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/merge-requests/{id}/accept": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Accepts and merge request and merges the respective users\nRequires user permission: user_merge_review",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "global"
                ],
                "summary": "Accept merge request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "request id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/name-change-requests": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Lists all open name change requests\nRequires user permission: name_change_review",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "global"
                ],
                "summary": "List name change requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/global.NameChangeRequest"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "ApiKeyAuth": [
                            "authorization"
                        ]
                    }
                ],
                "description": "Creates a name change request for the user. Needs to be reviewed by a moderator.\nRequires user permission: user_request_name_change",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "global"
                ],
                "summary": "Name Change Request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name to change to",
                        "name": "new_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/name-change-requests/{id}/accept": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Accepts a name change request from a user\nRequires user permission: name_change_review",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "global"
                ],
                "summary": "Accept name change request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "request id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "aredl.NameUser": {
            "type": "object",
            "properties": {
                "global_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "global.NameChangeRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "new_name": {
                    "type": "string"
                },
                "user": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string"
                        },
                        "name": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "util.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Perform actions as a user. It is also used to access endpoints that require user permissions.",
            "type": "apiKey",
            "name": "api-key",
            "in": "header"
        }
    }
}